+++ src/__tests__/pages/beta/teachers/lessons/[lessonSlug]/downloads.test.tsx
@@ -1,4 +1,4 @@
-import { act, renderHook, screen } from "@testing-library/react";
+import { renderHook, screen, act } from "@testing-library/react";
 import { GetServerSidePropsContext, PreviewData } from "next";
 import { useForm } from "react-hook-form";
 import userEvent from "@testing-library/user-event";
@@ -9,29 +9,16 @@ import waitForNextTick from "../../../../../__helpers__/waitForNextTick";
 import renderWithSeo from "../../../../../__helpers__/renderWithSeo";
 import { mockSeoResult } from "../../../../../__helpers__/cms";
 import renderWithProviders from "../../../../../__helpers__/renderWithProviders";
+import "../../../../../__helpers__/LocalStorageMock";
 import LessonDownloadsPage, {
   getServerSideProps,
   LessonDownloadsPageProps,
   URLParams,
 } from "../../../../../../pages/beta/teachers/key-stages/[keyStageSlug]/subjects/[subjectSlug]/units/[unitSlug]/lessons/[lessonSlug]/downloads";
-import { items } from "../../../../../../components/SchoolPicker/SchoolPicker.test";
-import useSchoolPicker from "../../../../../../components/SchoolPicker/useSchoolPicker";
+import useLocalStorageForDownloads from "../../../../../../components/DownloadComponents/hooks/useLocalStorageForDownloads";
 import teachersKeyStageSubjectUnitsLessonsDownloadsFixtures from "../../../../../../node-lib/curriculum-api/fixtures/teachersKeyStageSubjectUnitsLessonsDownloads.fixture";
 const props = {
   curriculumData: teachersKeyStageSubjectUnitsLessonsDownloadsFixtures(),
-};
-
-const setInputValue = jest.fn();
-const setSelectedValue = jest.fn();
-const setSelectedRadio = jest.fn();
-
-let useSchoolPickerReturnData = {
-  data: items,
-  error: null,
-  setInputValue: setInputValue,
-  inputValue: "dor",
-  setSelectedValue: setSelectedValue,
-  selectedValue: "dor",
 };
 
 const getDownloadResourcesExistenceData = {
@@ -40,14 +27,6 @@ const getDownloadResourcesExistenceData = {
     "worksheet-pdf": true,
   },
 };
-
-jest.mock(
-  "../../../../../../components/SchoolPicker/useSchoolPicker.tsx",
-  () => ({
-    __esModule: true,
-    default: () => useSchoolPickerReturnData,
-  })
-);
 
 jest.mock("next/dist/client/router", () => require("next-router-mock"));
 jest.mock(
@@ -227,6 +206,52 @@ describe("pages/beta/teachers/lessons/[lessonSlug]/downloads", () => {
     });
   });
 
+  describe("initial form values", () => {
+    it("gets email from local storage if available", async () => {
+      const { result } = renderHook(() => useLocalStorageForDownloads());
+
+      act(() => {
+        result.current.setEmailInLocalStorage("test@test.com");
+      });
+
+      const { getByLabelText, getByDisplayValue } = renderWithProviders(
+        <LessonDownloadsPage {...props} />
+      );
+
+      expect(getByLabelText("Email address")).toBeInTheDocument();
+      expect(getByDisplayValue("test@test.com")).toBeInTheDocument();
+    });
+
+    it.skip("gets school from local storage if available", async () => {
+      const { result } = renderHook(() => useLocalStorageForDownloads());
+
+      act(() => {
+        result.current.setSchoolInLocalStorage("Primary School");
+      });
+
+      const { getByDisplayValue } = renderWithProviders(
+        <LessonDownloadsPage {...props} />
+      );
+
+      expect(getByDisplayValue("Primary School")).toBeInTheDocument();
+    });
+
+    it("gets terms from local storage if available", async () => {
+      const { result } = renderHook(() => useLocalStorageForDownloads());
+
+      act(() => {
+        result.current.setTermsInLocalStorage("true");
+      });
+
+      const { getByLabelText } = renderWithProviders(
+        <LessonDownloadsPage {...props} />
+      );
+
+      const terms = getByLabelText("I accept terms and conditions (required)");
+      expect(terms).toBeChecked();
+    });
+  });
+
   describe("SEO", () => {
     it("renders the correct SEO details", async () => {
       const { seo } = renderWithSeo(<LessonDownloadsPage {...props} />);
@@ -266,89 +291,5 @@ describe("pages/beta/teachers/lessons/[lessonSlug]/downloads", () => {
         )
       ).rejects.toThrowError("No context.params");
     });
-  });
-});
-
-describe("School picker and radio buttons", () => {
-  it("clears school picker inputValue if radio button is clicked", async () => {
-    const { getByTestId } = renderWithProviders(
-      <LessonDownloadsPage {...props} />
-    );
-    const input = screen.getByTestId("search-autocomplete-input");
-
-    // initial value from mocked hook
-    expect(input).toHaveValue("dor");
-
-    const radio = getByTestId("radio-download");
-    const user = userEvent.setup();
-    await user.click(radio);
-    await user.tab();
-
-    // HACK: wait for next tick
-    await waitForNextTick();
-
-    expect(setInputValue).toBeCalledWith("");
-    expect(radio).toBeChecked();
-  });
-  /**
-   * @todo find a way to test that the radio is cleared when school is selected
-   */
-  it.skip("clears clicked radio if school is selected", async () => {
-    useSchoolPickerReturnData = {
-      data: items,
-      error: null,
-      setInputValue: setInputValue,
-      inputValue: "d",
-      setSelectedValue: setSelectedValue,
-      selectedValue: "",
-    };
-    const { getByTestId, rerender } = renderWithProviders(
-      <LessonDownloadsPage {...props} />
-    );
-
-    const radio = getByTestId("radio-download");
-    const user = userEvent.setup();
-    await user.click(radio);
-    await user.tab();
-
-    // HACK: wait for next tick
-    await waitForNextTick();
-
-    const input = screen.getByTestId("search-autocomplete-input");
-    await userEvent.type(input, "Dorothy");
-    await userEvent.type(input, "Dorot");
-
-    rerender(<LessonDownloadsPage {...props} />);
-
-    expect(setSelectedRadio).toBeCalledWith("");
-    // expect(radio).not.toBeChecked();
-  });
-  it.skip("clears clicked radio if school is selected", async () => {
-    const { getByTestId, rerender } = renderWithProviders(
-      <LessonDownloadsPage {...props} />
-    );
-
-    const radio = getByTestId("radio-download");
-    const user = userEvent.setup();
-    await user.click(radio);
-    await user.tab();
-
-    rerender(<LessonDownloadsPage {...props} />);
-
-    const useSchoolPickerHook = renderHook(() => useSchoolPicker());
-    const { setSelectedValue, setInputValue } =
-      useSchoolPickerHook.result.current;
-    act(() => {
-      setSelectedValue("anything");
-      setInputValue("anything");
-    });
-
-    const input = screen.getByTestId("search-autocomplete-input");
-    await userEvent.type(input, "Dorothy");
-    await userEvent.type(input, "Dorot");
-
-    rerender(<LessonDownloadsPage {...props} />);
-
-    expect(radio).not.toBeChecked();
   });
 });